added support for playlists

+++ utils/musicPlayer.js
@@ -2,8 +2,11 @@ const { EmbedBuilder, ButtonBuilder, ActionRowBuilder, ButtonStyle } = require("
 const wait = require("util").promisify(setTimeout);
 const logger = require("../utils/logger");
 
+let msg = null;
+
 module.exports = {
     trackStart: async (client, queue, track) => {
+        if (msg) msg.delete();
         //console.log(queue)
         //console.log(track)
         const channel = queue.metadata.channel;
@@ -37,8 +40,7 @@ module.exports = {
         .setFooter({ text: `Meme Cultist | Version ${require('../package.json').version}`, iconURL: client.user.displayAvatarURL({dynamic: true}) })
         .setTimestamp();
 
-        let msg = await channel.send({embeds: [player], components: [row]});
-        console.log(queue.tracks)
+        msg = await channel.send({embeds: [player], components: [row]});
 
         const interval = setInterval(async () => {
             if (!queue.node.isPlaying() || queue.node.isPaused() || track.isStream) return clearInterval(interval);
@@ -60,28 +62,26 @@ module.exports = {
                 if (!queue.node.isPaused()) {
                     logger.log("paused");
                     await queue.node.pause();
-                    // pause collector time
-                    await collector.resetTimer({ time: 60000 });
+                    await collector.resetTimer({ time: 300000 }); // 5 minutes to respond
                     player.setTitle(`⏸️ Song Paused`);
                     row.components[0].setLabel("Resume").setEmoji("▶️");
+                    row.components[1].setDisabled(true);
                     await i.update({embed: [player], components: [row]});
                 } else {
-                        logger.log("resumed");
-                        await queue.node.resume();
-                        await msg.delete();
-                        return await collector.stop();
+                    logger.log("resumed");
+                    await queue.node.resume();
+                    await msg.delete();
+                    return await collector.stop();
                 }
             } else if (i.customId === "skip") {
                 try {
-                    const success = queue.node.skip();
-                    if (success) {
-                        for (let i = 0; i < row.components.length; i++) {
-                            row.components[i].setDisabled(true);
-                        }
-                        await i.update({embed: [player], components: [row]});
-                        await msg.delete();
-                        await collector.stop();
-                    } 
+                    if (!queue.node.isPlaying() || queue.node.isPaused()) {
+                        // TODO: fix the bot creating two message when the user tries to skip while paused
+                        return await i.reply({ content: `Unpause before trying to skip. Too lazy to fix this bug for now.`, ephemeral: true });
+                    }
+                    await queue.node.skip();
+                    await msg.delete();
+                    await collector.stop();
                 } catch (e) {
                     logger.error(e);
                 }
@@ -99,18 +99,25 @@ module.exports = {
         collector.on('end', ( collected, reason ) => {
             logger.log(`Collected ${collected.size} interactions. Reason: ${reason}`);
             if (reason === "time") {
-                const reply = msg.reply({content: `Are you still there? I'll stop playing music in 30 seconds if you don't respond.`});
-                wait(30000).then(() => {
-                    if (!queue.node.isPlaying()) {
-                        queue.delete();
-                        msg.delete();
-                        reply.delete();
-                    } else {
-                        reply.delete();
-                    }
-                });
-            }
+                if (queue.node.isPaused()) {
+                    const reply = msg.reply(`Are you still there? Music will be stopped in 30 seconds if you don't respond.`);
+                    wait(30000).then(async () => {
+                        if (queue.node.isPaused()) {
+                            await queue.delete();
+                            await msg.delete();
+                            await reply.delete();
+                        } else {
+                            await reply.delete();
+                        }
+                    });
+                } else {
+                    if (msg) msg.delete();
+                }
+            } 
         });
+    },
+    trackEnd: async (client, queue, track) => {
+        if (msg) msg.delete();
     },
 
 };
